requirements.txt
paddleocr
paddlepaddle
python-docx
fpdf